rm extra Copy()s
