+++ core/services/eth/models.go
@@ -43,20 +43,12 @@ func NewHead(number *big.Int, blockHash common.Hash, parentHash common.Hash, tim
 	}
 }
 
-// Copy returns a deep copy, excluding Parent.
-func (h *Head) Copy() Head {
-	c := *h
-	c.EVMChainID = utils.New(h.EVMChainID)
-	c.BaseFeePerGas = utils.New(h.BaseFeePerGas)
-	return c
-}
-
 // EarliestInChain recurses through parents until it finds the earliest one
-func (h *Head) EarliestInChain() Head {
+func (h *Head) EarliestInChain() *Head {
 	for h.Parent != nil {
 		h = h.Parent
 	}
-	return h.Copy()
+	return h
 }
 
 // IsInChain returns true if the given hash matches the hash of a head in the chain
